requirements.txt
flask
flask-smorest
flask-cors
bs4
python-dotenv